+++ src/main.rs
@@ -36,7 +36,7 @@ fn cpu_cycle(mut emulator:ResMut<Emulator>){
     emulator.cpu_cycle();
 }
 
-pub fn get_nibble_from_byte(data: u32, i: u32) -> u8 {
+pub fn get_nibble_from_byte(data: u128, i: u32) -> u8 {
     let nib_idx = i + 1;
     dbg!(
         format!("{:#b}", data >> (32 - nib_idx * 4)),
